update spinquant quantization options to be general purposed pre-quantization (#5797)
Summary:
Pull Request resolved: https://github.com/pytorch/executorch/pull/5797

We decided to use the same quantization scheme and checkpoint format for QAT + LoRA. This PR updates related quantization cli options to be general purposed for pre-quantized checkpoints.

Reviewed By: mergennachin

Differential Revision: D63708762

fbshipit-source-id: 08d862d3331868616ff8dae5c5d86a2d36f26cd7
