+++ pytest.ini
@@ -43,7 +43,7 @@ addopts =
     --ignore=backends/xnnpack/test/ops/linear.py
     --ignore=backends/xnnpack/test/models/llama2_et_example.py
     # T200992559: Add torchao to ET as core dependency
-    --ignore=examples/models/llama2/tests/test_spinquant_transforms.py
+    --ignore=examples/models/llama2/tests/test_pre_quantization_transforms.py
     --ignore=exir/backend/test/demos
     --ignore=exir/backend/test/test_backends.py
     --ignore=exir/backend/test/test_backends_lifted.py
